push_swap_radix

+++ sources/operations.c
@@ -6,7 +6,7 @@
 /*   By: cabo-ram <cabo-ram@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/12 14:40:02 by cabo-ram          #+#    #+#             */
-/*   Updated: 2024/12/24 10:59:46 by cabo-ram         ###   ########.fr       */
+/*   Updated: 2025/01/07 15:55:40 by cabo-ram         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -43,19 +43,19 @@ void	push(char *type, t_stack *stack)
 	if (ft_strncmp(type, "pa", 3) == 0 && stack->sizeb > 0)
 	{
 		temp = stack->b[0];
-		ft_memmove(stack->a + 1, stack->a, sizeof(int) * stack->sizea);
+		ft_memmove(stack->a + 1, stack->a, sizeof(long) * stack->sizea);
 		stack->a[0] = temp;
 		stack->sizeb--;
-		ft_memmove(stack->b, stack->b + 1, sizeof(int) * stack->sizeb);
+		ft_memmove(stack->b, stack->b + 1, sizeof(long) * stack->sizeb);
 		stack->sizea++;
 	}
 	else if (ft_strncmp(type, "pb", 3) == 0 && stack->sizea > 0)
 	{
 		temp = stack->a[0];
-		ft_memmove(stack->b + 1, stack->b, sizeof(int) * stack->sizeb);
+		ft_memmove(stack->b + 1, stack->b, sizeof(long) * stack->sizeb);
 		stack->b[0] = temp;
 		stack->sizea--;
-		ft_memmove(stack->a, stack->a + 1, sizeof(int) * stack->sizea);
+		ft_memmove(stack->a, stack->a + 1, sizeof(long) * stack->sizea);
 		stack->sizeb++;
 	}
 	ft_putendl_fd(type, 1);
@@ -70,14 +70,14 @@ void	rotate(char *type, t_stack *stack)
 	if (ft_strncmp(type, "ra", 2) == 0 && stack->sizea > 1)
 	{
 		temp = stack->a[0];
-		ft_memmove(stack->a, stack->a + 1, sizeof(int) * (stack->sizea - 1));
+		ft_memmove(stack->a, stack->a + 1, sizeof(long) * (stack->sizea - 1));
 		stack->a[stack->sizea - 1] = temp;
 		ft_putendl_fd("ra", 1);
 	}
 	else if (ft_strncmp(type, "rb", 2) == 0 && stack->sizeb > 1)
 	{
 		temp = stack->b[0];
-		ft_memmove(stack->b, stack->b + 1, sizeof(int) * (stack->sizeb - 1));
+		ft_memmove(stack->b, stack->b + 1, sizeof(long) * (stack->sizeb - 1));
 		stack->b[stack->sizeb - 1] = temp;
 		ft_putendl_fd("rb", 1);
 	}
@@ -92,14 +92,14 @@ void	reverse_rotate(char *type, t_stack *stack)
 	if (ft_strncmp(type, "rra", 3) == 0 && stack->sizea > 1)
 	{
 		temp = stack->a[stack->sizea - 1];
-		ft_memmove(stack->a + 1, stack->a, sizeof(int) * (stack->sizea - 1));
+		ft_memmove(stack->a + 1, stack->a, sizeof(long) * (stack->sizea - 1));
 		stack->a[0] = temp;
 		ft_putendl_fd("rra", 1);
 	}
 	else if (ft_strncmp(type, "rrb", 3) == 0 && stack->sizeb > 1)
 	{
 		temp = stack->b[stack->sizeb - 1];
-		ft_memmove(stack->b + 1, stack->b, sizeof(int) * (stack->sizeb - 1));
+		ft_memmove(stack->b + 1, stack->b, sizeof(long) * (stack->sizeb - 1));
 		stack->b[0] = temp;
 		ft_putendl_fd("rrb", 1);
 	}
